# openapi/openapi.yaml
...
openapi: 3.1.0
info:
  title: SPA Development Reference API
  version: 0.1.0
  description: |
    Normative browser-facing HTTP contract of the SPA development reference
    application, a generic request and approval sample.

    All business endpoints require a Bearer access token. In the AWS
    deployment (`APP_MODE=aws`) it is a JWT access token issued by Amazon
    Cognito and the application role of the caller (Requester, Approver, or
    Administrator) is derived from the `cognito:groups` claim of the validated
    token. In the local demo (`APP_MODE=local`) it is one of three fixed demo
    tokens and the backend maps each to one role. In both modes the roles are
    derived by the server, and each operation states the roles it requires.
    A caller that lacks the required role, or that is not permitted to use the
    addressed resource, receives `403 FORBIDDEN`. The Administrator role is not
    a superuser.

    Evaluation order for a business operation: authentication, coarse role,
    structural validation (path, query, content type, body), resource lookup,
    resource authorization, version, state. The first failing step determines
    the response; in particular, when a supplied `version` is stale and the
    current state is also invalid, the response is `409 CONCURRENCY_CONFLICT`.
    An unknown path is `404 ROUTE_NOT_FOUND` and an unsupported method on a known
    path is `405 METHOD_NOT_ALLOWED`; both are Problem responses. The detailed
...
servers:
...
security:
  - bearerAuth: []
paths:
...
  /api/requests/{requestId}/attachments:
    parameters:
      - $ref: '#/components/parameters/RequestId'
    get:
      operationId: listAttachments
      tags: [Attachments]
      summary: List the attachments of a request
      description: |
        Lists the attachment metadata of the request, so that attachment
        identifiers can be rediscovered at any time.

        Requester: attachments of an own request, in any request state.
        Approver: attachments of a request in `SUBMITTED` state. Administrator:
        attachments of any request. Otherwise `403 FORBIDDEN`; a caller with none
        of these roles receives `403` before the existence of the request is
        revealed. A request that does not exist is `404 REQUEST_NOT_FOUND`.

        Order: `createdAt` ascending, then `id` ascending. Pagination is applied
        after the authorization check. The cursor is bound to this endpoint and to
        the `requestId`; using it with another request is a
        `400 VALIDATION_ERROR`.
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Cursor'
      responses:
        '200':
          description: Attachment metadata of the request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AttachmentList'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
        '503':
          $ref: '#/components/responses/IdentityProviderUnavailable'
    post:
      operationId: uploadAttachment
      tags: [Attachments]
      summary: Attach a file to a request
      description: |
        Requires the Requester role and ownership of the request (otherwise
        `403 FORBIDDEN`). The request must be in `DRAFT` state; otherwise
        `409 REQUEST_INVALID_STATE`. The state is checked again when the
        attachment is finalized.

        Accepted media types: `application/pdf`, `image/png`, `image/jpeg`, and
        `text/plain`. The server validates the content against the declared
        media type, and the file name extension must match the media type.
        A mismatch or an unsupported type is `415 UNSUPPORTED_MEDIA_TYPE`.

        The default maximum size is 10 MiB and is configurable by deployment up
        to 100 MiB; a larger file is `413 PAYLOAD_TOO_LARGE`. A zero-byte file is
        rejected as `400 VALIDATION_ERROR`. File names that are empty, longer than
        255 characters, or that contain `/`, `\`, or control characters are also
        rejected as `400 VALIDATION_ERROR`.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
                  description: The file content.
      responses:
        '201':
          description: The attachment was stored.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Attachment'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '413':
          $ref: '#/components/responses/PayloadTooLarge'
        '415':
          $ref: '#/components/responses/UnsupportedMediaType'
        '500':
          $ref: '#/components/responses/InternalError'
        '503':
          $ref: '#/components/responses/DependencyUnavailable'
  /api/requests/{requestId}/attachments/{attachmentId}:
    parameters:
      - $ref: '#/components/parameters/RequestId'
      - $ref: '#/components/parameters/AttachmentId'
    get:
      operationId: downloadAttachment
      tags: [Attachments]
      summary: Download an attachment
      description: |
        Requester: an attachment of an own request, in any request state.
        Approver: an attachment of a request in `SUBMITTED` state.
        Administrator: any attachment. Otherwise `403 FORBIDDEN`. An attachment
        that does not belong to the `requestId` in the path is
        `404 ATTACHMENT_NOT_FOUND`.
      responses:
        '200':
          description: The attachment content, always as a download (never inline).
          headers:
            Content-Disposition:
              description: |
                `attachment` with a safe ASCII `filename` fallback and an
                RFC 8187 `filename*` carrying the stored file name.
              schema:
                type: string
            Content-Length:
              description: Stored attachment size in bytes.
              schema:
                type: integer
                minimum: 1
            X-Content-Type-Options:
              description: Always `nosniff`.
              schema:
                type: string
                enum: [nosniff]
            Cache-Control:
              description: Always `private, no-store`.
              schema:
                type: string
                enum: ['private, no-store']
          content:
            application/pdf:
              schema:
                type: string
                format: binary
            image/png:
              schema:
                type: string
                format: binary
            image/jpeg:
              schema:
                type: string
                format: binary
            text/plain:
              schema:
                type: string
                format: binary
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
        '503':
          $ref: '#/components/responses/DependencyUnavailable'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        AWS deployment: Amazon Cognito access token only (an ID token is not
        accepted as authorization evidence). The BFF validates the token and
        derives the application roles from its `cognito:groups` claim.
        Local demo (`APP_MODE=local`): one of the fixed, non-secret demo tokens,
        which the AWS deployment never accepts. In both modes role values
        supplied by the browser are never used.
  parameters:
    RequestId:
      name: requestId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    AttachmentId:
      name: attachmentId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    Limit:
      name: limit
      in: query
      required: false
      description: Maximum number of items in the page.
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 50
    Cursor:
      name: cursor
      in: query
      required: false
      description: |
        Opaque cursor taken from `nextCursor` of the previous page. Clients do
        not interpret its content. Cursors do not expire. A cursor may be bound to
        endpoint-specific filters and resource identifiers: the audit list binds
        it to the `requestId` filter, and the attachment list binds it to the
        `requestId` path value. A malformed cursor, or a cursor used with a
        different endpoint or a different bound value, is a
        `400 VALIDATION_ERROR`. Pages are not
        taken from a database snapshot: concurrent changes between page requests
        can move or skip items.
      schema:
        type: string
        minLength: 1
  responses:
    BadRequest:
      description: Validation failure (`VALIDATION_ERROR`).
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          example:
            title: Validation error
            status: 400
            detail: The title must be between 1 and 200 characters.
            code: VALIDATION_ERROR
            traceId: example-trace-id
    Unauthorized:
      description: Authentication is required or the token is invalid (`AUTHENTICATION_REQUIRED`).
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          example:
            title: Authentication required
            status: 401
            code: AUTHENTICATION_REQUIRED
            traceId: example-trace-id
    Forbidden:
      description: The caller is not permitted to perform the operation (`FORBIDDEN`).
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          example:
            title: Forbidden
            status: 403
            code: FORBIDDEN
            traceId: example-trace-id
    NotFound:
      description: The request or attachment does not exist (`REQUEST_NOT_FOUND` or `ATTACHMENT_NOT_FOUND`).
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          example:
            title: Request not found
            status: 404
            code: REQUEST_NOT_FOUND
            traceId: example-trace-id
    Conflict:
      description: The supplied version is stale (`CONCURRENCY_CONFLICT`, which takes precedence) or, with a matching version, the request is not in a valid state for the operation (`REQUEST_INVALID_STATE`).
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          example:
            title: Invalid state transition
            status: 409
            detail: The request is not in a state that can be approved.
            code: REQUEST_INVALID_STATE
            traceId: example-trace-id
...
    PayloadTooLarge:
      description: The uploaded file exceeds the configured maximum size (`PAYLOAD_TOO_LARGE`).
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          example:
            title: Payload too large
            status: 413
            code: PAYLOAD_TOO_LARGE
            traceId: example-trace-id
    UnsupportedMediaType:
      description: The media type of the request body or uploaded file is not supported, or the file content or extension does not match it (`UNSUPPORTED_MEDIA_TYPE`).
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          example:
            title: Unsupported media type
            status: 415
            code: UNSUPPORTED_MEDIA_TYPE
            traceId: example-trace-id
    IdentityProviderUnavailable:
      description: The identity provider needed to verify the token or the caller's email is unavailable (`IDENTITY_PROVIDER_UNAVAILABLE`).
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          example:
            title: Identity provider unavailable
            status: 503
            code: IDENTITY_PROVIDER_UNAVAILABLE
            traceId: example-trace-id
    DependencyUnavailable:
      description: |
        A required dependency is unavailable: either the identity provider
        (`IDENTITY_PROVIDER_UNAVAILABLE`) or object storage
        (`OBJECT_STORAGE_UNAVAILABLE`).
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          examples:
            identityProvider:
              value:
                title: Identity provider unavailable
                status: 503
                code: IDENTITY_PROVIDER_UNAVAILABLE
                traceId: example-trace-id
            objectStorage:
              value:
                title: Object storage unavailable
                status: 503
                code: OBJECT_STORAGE_UNAVAILABLE
                traceId: example-trace-id
    InternalError:
      description: Unexpected implementation, database, or unclassified infrastructure failure (`INTERNAL_ERROR`).
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          example:
            title: Internal error
            status: 500
            code: INTERNAL_ERROR
            traceId: example-trace-id
  schemas:
...
    Attachment:
      type: object
      required: [id, requestId, fileName, mediaType, sizeBytes, uploadedBy, createdAt]
      properties:
        id:
          type: string
          format: uuid
        requestId:
          type: string
          format: uuid
        fileName:
          type: string
          minLength: 1
          maxLength: 255
        mediaType:
          type: string
          minLength: 1
          maxLength: 255
          description: One of `application/pdf`, `image/png`, `image/jpeg`, or `text/plain`.
        sizeBytes:
          type: integer
          minimum: 1
        uploadedBy:
          type: string
          minLength: 1
        createdAt:
          type: string
          format: date-time
    AttachmentList:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Attachment'
        nextCursor:
          type: string
          minLength: 1
          description: Cursor for the next page. Omitted when there is no next page.
...
    ErrorCode:
      description: Stable machine-readable application error identifier.
      type: string
      enum:
        - VALIDATION_ERROR
        - AUTHENTICATION_REQUIRED
        - FORBIDDEN
        - REQUEST_NOT_FOUND
        - ATTACHMENT_NOT_FOUND
        - ROUTE_NOT_FOUND
        - METHOD_NOT_ALLOWED
        - REQUEST_INVALID_STATE
        - CONCURRENCY_CONFLICT
        - PAYLOAD_TOO_LARGE
        - UNSUPPORTED_MEDIA_TYPE
        - IDENTITY_PROVIDER_UNAVAILABLE
        - OBJECT_STORAGE_UNAVAILABLE
        - INTERNAL_ERROR
    Problem:
      description: |
        Problem Details (`application/problem+json`) with stable application
        extension fields. Status mapping: `VALIDATION_ERROR` 400,
        `AUTHENTICATION_REQUIRED` 401, `FORBIDDEN` 403, `REQUEST_NOT_FOUND` and
        `ATTACHMENT_NOT_FOUND` 404, `ROUTE_NOT_FOUND` 404,
        `METHOD_NOT_ALLOWED` 405, `REQUEST_INVALID_STATE` and
        `CONCURRENCY_CONFLICT` 409, `PAYLOAD_TOO_LARGE` 413,
        `UNSUPPORTED_MEDIA_TYPE` 415, `IDENTITY_PROVIDER_UNAVAILABLE` and
        `OBJECT_STORAGE_UNAVAILABLE` 503, `INTERNAL_ERROR` 500.
      type: object
      required: [title, status, code, traceId]
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string
          format: uri-reference
        code:
          $ref: '#/components/schemas/ErrorCode'
        traceId:
          type: string